Synchronous render-readback-display implementation
This makes readback thread wait until the display thread has drawn the frame
buffer and swapped buffers on the window, and only after that unblock the main
thread to let it swap buffers on the Pbuffer and release control to the
application.

Therefore, there is no pipelining of work, so performance suffers. On a
positive side, this should avoid crashes or hangs when the window destroyed
when D::work is still attempting to draw on it.