Try to save power by sleeping before glMapBuffer
The expectation here is that glMapBuffer spends most of the time busy waiting
for pixel data arrival, so try to sleep for most of the waiting time instead.
Since we can't know how much time glMapBuffer time is going to take, use the
time from the previous frame and sleep for a fixed fraction of that (90% by
default).